backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-multipart==0.0.6

# HTTP Client for AI Service
httpx==0.26.0

# Phase 2 placeholders (not used yet)
# langchain==0.1.0
# langgraph==0.0.20
# qdrant-client==1.7.0
# openai==1.10.0
